+++ iceland.html
@@ -51,14 +51,14 @@
             </article>
             <article class="famousCitizens countryArticles">
                 <h2><u>People you may know</u></h2>
-                <div class="cities">
+                <!-- <div class="cities">
                     <img src="https://cdn.britannica.com/43/103043-050-1C637E5D/Bjork-opening-ceremony-Athens-2004-Olympic-Games-2004.jpg" alt="Singer Bjork">
                     <p>Björk (singer-songwriter)</p> 
                     <img src="https://tmssl.akamaized.net/images/foto/normal/gylfi-sigurdsson-everton-1538385375-18000.jpg" alt="Gylfi Sigurðsson (Icelandic soccer player)">
                     <p>Gylfi Sigurðsson (Icelandic soccer player) </p>
                     <img src="https://guidetoiceland.imgix.net/343356/x/0/omam-jpg?w=860&h=480&fit=crop&auto=format%2C%20compress&dpr=2&ixlib=react-8.6.4" alt="Of Monsters and Men (Rock band)">
                     <p>Of Monsters and Men (Rock band)</p>
-                </div>
+                </div> -->
             </article>
             
         </section>
