+++ scripts/Iceland/celebrities/Celebrities.js
@@ -1,0 +1,6 @@
+export const Celebrities = (celebrity) => {
+    return `
+        <img src=${celebrity.image}>
+        <p>${celebrity.name}</p> 
+    `
+}
