--- scripts/Japan/celebrities/Celebrities.js
@@ -1,6 +1,0 @@
-export const Celebrities = (celebrity) => {
-    return `
-        <img src=${celebrity.image}>
-        <p>${celebrity.name}</p> 
-    `
-}
